adjusting data objects
